# src/internals/controller/order.go
package controller

import (
	"encoding/json"
	"errors"
	"net/http"
	"strconv"
	"strings"

	"github.com/audricimanuel/awb-stock-allocation/src/internals/dto"
	"github.com/audricimanuel/awb-stock-allocation/src/internals/service"
	"github.com/audricimanuel/awb-stock-allocation/src/model"
	e "github.com/audricimanuel/awb-stock-allocation/utils/errors"
	"github.com/audricimanuel/awb-stock-allocation/utils/httputils"
	"github.com/audricimanuel/errorutils"
	"github.com/go-chi/chi/v5"
	"github.com/sirupsen/logrus"
)

// TODO: the interface and implementation under here
type (
	OrderController interface {
		CreateOrder(w http.ResponseWriter, r *http.Request)
		GetOrderById(w http.ResponseWriter, r *http.Request)
		UpdateOrderStatus(w http.ResponseWriter, r *http.Request)
		GetOrders(w http.ResponseWriter, r *http.Request)
	}

	OrderControllerImpl struct {
		orderService service.OrderService
		logger       *logrus.Logger
	}
)

func NewOrderController(orderService service.OrderService, logger *logrus.Logger) OrderController {
	return &OrderControllerImpl{
		orderService: orderService,
		logger:       logger,
	}
}

// @Tags			Order
// @Summary		Create Order
// @Description	"Create an Order"
// @Accept			json
// @Produce		json
// @Param			payload	body		dto.CreateOrderRequest	true	"Order request payload"
// @Success		200		{object}	httputils.BaseResponse
// @Router			/orders [post]
func (a *OrderControllerImpl) CreateOrder(w http.ResponseWriter, r *http.Request) {
	var req *dto.CreateOrderRequest

	err := json.NewDecoder(r.Body).Decode(&req)
	if err != nil {
		a.logger.WithError(err).Warn("failed to decode request body")
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorInternalServer, nil)
		return
	}

	if req.AWBNumber == "" {
		err = e.ErrOrderAWBRequired

		a.logger.WithError(err).Warn(err.Error())
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
		return
	}

	if req.Sender == "" {
		err = e.ErrOrderSenderRequired

		a.logger.WithError(err).Warn(err.Error())
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
		return
	}

	if req.Receiver == "" {
		err = e.ErrOrderReceiverRequired

		a.logger.WithError(err).Warn(err.Error())
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
		return
	}

	if req.TotalWeight == 0 {
		err = e.ErrOrderTotalWeightRequired

		a.logger.WithError(err).Warn(err.Error())
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
		return
	}

	order := &model.Order{
		AWBNumber:   req.AWBNumber,
		Sender:      req.Sender,
		Receiver:    req.Receiver,
		TotalWeight: req.TotalWeight,
	}

	a.logger.WithFields(logrus.Fields{
		"awb_number":   order.AWBNumber,
		"sender":       order.Sender,
		"receiver":     order.Receiver,
		"total_weight": order.TotalWeight,
	}).Info("creating order")

	orderModel, err := a.orderService.CreateOrder(order)
	if err != nil {
		a.logger.WithError(err).Warn(err.Error())

		if err == e.ErrAWBHasBeenUsed {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
			return
		}

		if err == e.ErrAWBInvalid {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
			return
		}

		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorInternalServer.CustomMessage(err.Error()), nil)
		return
	}

	resp := &dto.CreateOrderResponse{
		ID:          orderModel.ID,
		AWBNumber:   orderModel.AWBNumber,
		Sender:      orderModel.Sender,
		Receiver:    orderModel.Receiver,
		TotalWeight: orderModel.TotalWeight,
		TotalPrice:  orderModel.TotalPrice,
		Status:      orderModel.Status,
	}

	a.logger.WithField("awb_number", orderModel.ID).Info("order created successfully")

	httputils.MapBaseResponse(w, r, resp, nil, nil)
}

// @Tags			Order
// @Summary		Update Order Status
// @Description	"Update an Order Status"
//
// @Param			id		path	int								true	"Order ID"
//
// @Param			payload	body	dto.UpdateOrderStatusRequest	true	"Order status payload. Allowed values: 1=PENDING, 2=CONFIRM, 3=SHIPPED, 4=COMPLETED, 5=CANCELLED"
//
// @Accept			json
// @Produce		json
// @Success		200	{object}	httputils.BaseResponse
// @Router			/orders/{id}/status [put]
func (a *OrderControllerImpl) UpdateOrderStatus(w http.ResponseWriter, r *http.Request) {
	var req *dto.UpdateOrderStatusRequest

	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		err = errors.New("invalid order id")

		a.logger.WithError(err).Error(err.Error())
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorNotFound.CustomMessage(err.Error()), nil)
		return
	}

	err = json.NewDecoder(r.Body).Decode(&req)
	if err != nil {
		a.logger.WithError(err).Error(err.Error())
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorInternalServer.CustomMessage(err.Error()), nil)
		return
	}

	a.logger.WithFields(logrus.Fields{
		"id":     id,
		"status": req.Status,
	}).Info("updating order status")

	order, err := a.orderService.UpdateOrderStatus(id, req.Status)
	if err != nil {
		if err == e.ErrOrderStatusInvalid {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorInvalidPayload.CustomMessage(err.Error()), nil)
			return
		}

		if err == e.ErrOrderNotFound {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorNotFound.CustomMessage(err.Error()), nil)
			return
		}

		if err == e.ErrOrderStatusPendingValidation {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
			return
		}

		if err == e.ErrOrderStatusConfirmValidation {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
			return
		}

		if err == e.ErrOrderStatusShippedValidation {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
			return
		}

		if err == e.ErrOrderStatusFinal {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorBadRequest.CustomMessage(err.Error()), nil)
			return
		}

		a.logger.WithError(err).Error(err.Error())
		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorInternalServer.CustomMessage(err.Error()), nil)
		return
	}

	resp := &dto.CreateOrderResponse{
		ID:          order.ID,
		AWBNumber:   order.AWBNumber,
		Sender:      order.Sender,
		Receiver:    order.Receiver,
		TotalWeight: order.TotalWeight,
		TotalPrice:  order.TotalPrice,
		Status:      order.Status,
	}

	a.logger.WithField("awb_number", order.ID).Info("order status updated successfully")
	httputils.MapBaseResponse(w, r, resp, nil, nil)
}

// @Tags			Order
// @Summary		Get Order by Order ID
// @Description	"Get an Order by Order ID"
// @Param			id	path	int	true	"Order ID"
// @Accept			json
// @Produce		json
// @Success		200	{object}	httputils.BaseResponse
// @Router			/orders/{id} [get]
func (a *OrderControllerImpl) GetOrderById(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		err = errors.New("invalid order id")

		a.logger.WithFields(logrus.Fields{
			"id_param": idStr,
			"error":    err.Error(),
		}).Warn("invalid order id")

		httputils.MapBaseResponse(w, r, nil, err, nil)
		return
	}

	order, err := a.orderService.GetOrderById(id)
	if err != nil {
		a.logger.WithFields(logrus.Fields{
			"order_id": id,
			"error":    err.Error(),
		}).Warn("failed to fetch order")

		if err == e.ErrOrderNotFound {
			httputils.MapBaseResponse(w, r, nil, errorutils.ErrorNotFound.CustomMessage(err.Error()), nil)
			return
		}

		httputils.MapBaseResponse(w, r, nil, errorutils.ErrorInternalServer.CustomMessage(err.Error()), nil)
		return
	}

	resp := &dto.CreateOrderResponse{
		ID:          order.ID,
		AWBNumber:   order.AWBNumber,
		Sender:      order.Sender,
		Receiver:    order.Receiver,
		TotalWeight: order.TotalWeight,
		TotalPrice:  order.TotalPrice,
		Status:      order.Status,
	}

	a.logger.WithFields(logrus.Fields{
		"order_id":     order.ID,
		"awb_number":   order.AWBNumber,
		"sender":       order.Sender,
		"receiver":     order.Receiver,
		"total_weight": order.TotalWeight,
	}).Info("successfully fetched order")

	httputils.MapBaseResponse(w, r, resp, nil, nil)
}

// @Tags			Order
// @Summary		Get Order List
// @Description	"Get an Order List"
// @Accept			json
// @Produce		json
// @Param			page	query		int		false	"Page number"						example(1)
// @Param			awb		query		string	false	"Search by AWB number substring"	example("001-10000001")
// @Success		200		{object}	httputils.BaseResponse
// @Router			/orders [get]
func (a *OrderControllerImpl) GetOrders(w http.ResponseWriter, r *http.Request) {
	pageStr := r.URL.Query().Get("page")
	awb := strings.TrimSpace(r.URL.Query().Get("awb"))

	page := 1
	if pageStr != "" {
		if p, err := strconv.Atoi(pageStr); err == nil && p > 0 {
			page = p
		}
	}

	orders, _ := a.orderService.GetOrders(page, awb)

	var resp []*dto.CreateOrderResponse

	for _, order := range orders {
		r := &dto.CreateOrderResponse{
			ID:          order.ID,
			AWBNumber:   order.AWBNumber,
			Sender:      order.Sender,
			Receiver:    order.Receiver,
			TotalWeight: order.TotalWeight,
			TotalPrice:  order.TotalPrice,
			Status:      order.Status,
		}

		resp = append(resp, r)
	}

	meta := httputils.SetBaseMeta(page, 5, len(orders))

	a.logger.Info("get order list successful")
	httputils.MapBaseResponse(w, r, resp, nil, &meta)
}